refactor(desktop): extract reusable Pocket voice primitives
