+++ crates/buzz-voice/src/lib.rs
@@ -1,0 +1,3 @@
+//! Reusable local voice primitives for Buzz.
+
+pub mod pocket;
